Moving the timeseries maps to QtCharts.
